backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pillow==11.2.1
numpy==2.2.5
torch==2.6.0
transformers==4.51.3
open-clip-torch==2.29.0
requests==2.32.3
aiofiles==24.1.0
python-dotenv==1.1.0
tqdm==4.67.1
